+++ packages/google_mlkit_text_recognition/ios/Classes/GoogleMlKitTextRecognitionPlugin.m
@@ -35,6 +35,13 @@
     [registrar addMethodCallDelegate:instance channel:channel];
 }
 
+- (id)init {
+    self = [super init];
+    if (self)
+        instances = [NSMutableDictionary dictionary];
+    return  self;
+}
+
 - (void)handleMethodCall:(FlutterMethodCall *)call result:(FlutterResult)result {
     if ([call.method isEqualToString:startTextRecognizer]) {
         [self handleDetection:call result:result];
@@ -50,7 +57,7 @@
 - (MLKTextRecognizer*)initialize:(FlutterMethodCall *)call {
     NSNumber *scriptValue = call.arguments[@"script"];
     switch(scriptValue.intValue) {
-        case 0 : {
+        default : {
             MLKTextRecognizerOptions *latinOptions = [[MLKTextRecognizerOptions alloc] init];
             return [MLKTextRecognizer textRecognizerWithOptions:latinOptions];
         }
@@ -78,8 +85,6 @@
             return [MLKTextRecognizer textRecognizerWithOptions:koreanOptions];
         }
 #endif
-        default:
-            return NULL;
     }
 }
 
@@ -116,8 +121,8 @@
                     frame:block.frame
                 languages:block.recognizedLanguages
                      text:block.text
-               confidence:NULL
-                    angle:NULL];
+               confidence:[NSNull null]
+                    angle:[NSNull null]];
             
             NSMutableArray *textLines = [NSMutableArray array];
             for (MLKTextLine *line in block.lines) {
@@ -128,8 +133,9 @@
                         frame:line.frame
                     languages:line.recognizedLanguages
                          text:line.text
-                   confidence:line.confidence
-                        angle:line.angle];
+                   confidence:[NSNull null] // TODO: replace with actual value once it is supported by Google's native API
+                        angle:[NSNull null]]; // TODO: replace with actual value once it is supported by Google's native API
+                // API: https://developers.google.com/ml-kit/reference/ios/mlkittextrecognitioncommon/api/reference/Classes/MLKTextLine
                 
                 NSMutableArray *elementsData = [NSMutableArray array];
                 for (MLKTextElement *element in line.elements) {
@@ -138,11 +144,14 @@
                     [self addData:elementData
                      cornerPoints:element.cornerPoints
                             frame:element.frame
-                        languages:NULL
+                        languages:element.recognizedLanguages
                              text:element.text
-                       confidence:element.confidence
-                            angle:element.angle];
+                       confidence:[NSNull null] // TODO: replace with actual value once it is supported by Google's native API
+                            angle:[NSNull null]]; // TODO: replace with actual value once it is supported by Google's native API
+                    // API: https://developers.google.com/ml-kit/reference/ios/mlkittextrecognitioncommon/api/reference/Classes/MLKTextElement
                     
+                    // TODO: add when Google's native API supports it
+                    elementData[@"symbols"] = @[];
                     [elementsData addObject:elementData];
                 }